benchmarks/ml-pipeline/ml-vehicle/requirements.txt
# setup modules
gevent==1.4.0
flask==1.0.2
# packages for numerics
numpy == 1.18.4
scikit-learn == 0.22.2
scipy == 1.4.1
# packages for image processing & common utils
Pillow == 7.1.2
# packages for database & object storage
cloudant == 2.14.0
minio == 5.0.10
# ml
mxnet
gluncv
